moved section stuff to package; hiding registry impl
