+++ spring-auto-restdocs-core/src/main/java/capital/scalable/restdocs/AutoDocumentation.java
@@ -19,11 +19,11 @@ package capital.scalable.restdocs;
 import capital.scalable.restdocs.misc.AuthorizationSnippet;
 import capital.scalable.restdocs.misc.DescriptionSnippet;
 import capital.scalable.restdocs.misc.MethodAndPathSnippet;
-import capital.scalable.restdocs.misc.SectionBuilder;
 import capital.scalable.restdocs.payload.JacksonRequestFieldSnippet;
 import capital.scalable.restdocs.payload.JacksonResponseFieldSnippet;
 import capital.scalable.restdocs.request.PathParametersSnippet;
 import capital.scalable.restdocs.request.RequestParametersSnippet;
+import capital.scalable.restdocs.section.SectionBuilder;
 import org.springframework.restdocs.snippet.Snippet;
 
 public abstract class AutoDocumentation {
