+++ spring-auto-restdocs-core/src/main/java/capital/scalable/restdocs/misc/MethodAndPathSnippet.java
@@ -18,7 +18,7 @@ package capital.scalable.restdocs.misc;
 
 import static capital.scalable.restdocs.OperationAttributeHelper.getRequestMethod;
 import static capital.scalable.restdocs.OperationAttributeHelper.getRequestPattern;
-import static capital.scalable.restdocs.misc.SnippetRegistry.METHOD_PATH;
+import static capital.scalable.restdocs.SnippetRegistry.METHOD_PATH;
 
 import java.util.HashMap;
 import java.util.Map;
