+++ spring-auto-restdocs-core/src/main/java/capital/scalable/restdocs/payload/JacksonRequestFieldSnippet.java
@@ -16,7 +16,7 @@
 
 package capital.scalable.restdocs.payload;
 
-import static capital.scalable.restdocs.misc.SnippetRegistry.REQUEST_FIELDS;
+import static capital.scalable.restdocs.SnippetRegistry.REQUEST_FIELDS;
 
 import java.lang.reflect.GenericArrayType;
 import java.lang.reflect.Type;
