+++ spring-auto-restdocs-core/src/main/java/capital/scalable/restdocs/payload/JacksonResponseFieldSnippet.java
@@ -16,7 +16,7 @@
 
 package capital.scalable.restdocs.payload;
 
-import static capital.scalable.restdocs.misc.SnippetRegistry.RESPONSE_FIELDS;
+import static capital.scalable.restdocs.SnippetRegistry.RESPONSE_FIELDS;
 
 import java.lang.reflect.GenericArrayType;
 import java.lang.reflect.Type;
