+++ spring-auto-restdocs-core/src/main/java/capital/scalable/restdocs/request/PathParametersSnippet.java
@@ -16,7 +16,7 @@
 
 package capital.scalable.restdocs.request;
 
-import static capital.scalable.restdocs.misc.SnippetRegistry.PATH_PARAMETERS;
+import static capital.scalable.restdocs.SnippetRegistry.PATH_PARAMETERS;
 
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.annotation.PathVariable;
